search, comments & authentication

+++ pswapp/views.py
@@ -1,3 +1,4 @@
+from urllib.parse import quote_plus
 from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect
 from .models import Post
 from .forms import NewPostForm
@@ -10,6 +11,9 @@ def home(request):
 # Blog Posts
 def posts_home(request):
     queryset = Post.objects.all().order_by("-timestamp")
+    query = request.GET.get("q")
+    if query:
+        queryset = queryset.filter(title__icontains=query)
     context = {
         "object_list": queryset,
         "title": "List"
@@ -17,6 +21,9 @@ def posts_home(request):
     return render(request, 'blog.html', context)
 
 def post_create(request):
+    if not request.staff or not request.user.is_superuser:
+        raise Http404
+
     new_post = NewPostForm(request.POST or None, request.FILES or None)
     if new_post.is_valid():
         instance = new_post.save(commit=False)
@@ -30,13 +37,18 @@ def post_create(request):
 
 def post_detail(request, id):
     post = get_object_or_404(Post, id=id)
+    share_string = quote_plus(post.content)
     context = {
         "title": post.title,
         "post": post,
+        "share_string": share_string,
     }
     return render(request, 'blog_post.html', context)
 
 def post_update(request, id):
+    if not request.staff or not request.user.is_superuser:
+        raise Http404
+
     post = get_object_or_404(Post, id=id)
     new_post = NewPostForm(request.POST or None, request.FILES or None, instance=post)
     if new_post.is_valid():
@@ -54,6 +66,9 @@ def post_update(request, id):
 
 
 def post_delete(request, id):
+    if not request.staff or not request.user.is_superuser:
+        raise Http404
+
     post = get_object_or_404(Post, id=id)
     post.delete()
     messages.success(request, "Successfully Deleted Post")
